requirements.txt
# 텔레그램 Claude 비서 봇
python-telegram-bot==21.6
google-genai>=0.8.0
python-dotenv>=1.0.0

# Reddit 블로그 봇
requests>=2.31.0
schedule>=1.2.0

# YouTube 요약 봇
youtube-transcript-api>=0.6.2
